Rename DomainEventAssertions class.
